[Ingest Manager] Move cache functions to from registry to archive (#82871)
## Summary

Moving the memory store functions to `archive/cache.ts` to better express their role. 

They are archive-related functions. The registry is just one possible source of an archive/assets. Also considered moving to `assets`, but can always come back to that.

`deletePackageCache` undoes side-effects from `unpackArchiveToCache` so put them in the same file
