+++ x-pack/plugins/ingest_manager/server/services/epm/packages/remove.ts
@@ -18,7 +18,8 @@ import { deletePipeline } from '../elasticsearch/ingest_pipeline/';
 import { installIndexPatterns } from '../kibana/index_pattern/install';
 import { deleteTransforms } from '../elasticsearch/transform/remove';
 import { packagePolicyService, appContextService } from '../..';
-import { splitPkgKey, deletePackageCache } from '../registry';
+import { splitPkgKey } from '../registry';
+import { deletePackageCache } from '../archive';
 
 export async function removeInstallation(options: {
   savedObjectsClient: SavedObjectsClientContract;
